refactor: infra member package 생성
